feature-erms-5004
as of ERMS-5004, price items may be added or removed without reloading of page
